# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FitnessApp.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Exercise": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exercise"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Exercise/{id}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Exercise/search": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/muscleGroup/{group}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/equipment/{equipment}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "equipment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/category/{category}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/difficulty/{level}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Exercise/type/{type}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Exercise/update/{id}": {
      "post": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Exercise/{id}/aimodel": {
      "put": {
        "tags": [
          "Exercise"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIModelReferenceUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AIModelReferenceUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AIModelReferenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIModelReferenceUpdate": {
        "type": "object",
        "properties": {
          "aiModelReference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Exercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "muscleGroup": {
            "type": "string",
            "nullable": true
          },
          "equipment": {
            "type": "string",
            "nullable": true
          },
          "difficultyLevel": {
            "type": "string",
            "nullable": true
          },
          "instructionSteps": {
            "type": "string",
            "nullable": true
          },
          "videoUrl": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "aiModelReference": {
            "type": "string",
            "nullable": true
          },
          "workoutExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExerciseDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "muscleGroup": {
            "type": "string",
            "nullable": true
          },
          "equipment": {
            "type": "string",
            "nullable": true
          },
          "difficultyLevel": {
            "type": "string",
            "nullable": true
          },
          "instructionSteps": {
            "type": "string",
            "nullable": true
          },
          "videoUrl": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "aiModelReference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WorkoutDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "workoutPlanId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutExercise"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "workoutDayId": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "repsPerSet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderInWorkout": {
            "type": "integer",
            "format": "int32"
          },
          "restBetweenSets": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "workoutDay": {
            "$ref": "#/components/schemas/WorkoutDay"
          },
          "exercise": {
            "$ref": "#/components/schemas/Exercise"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}